fix(pc-port): Drop -fpermissive, which inflated the compile metric
tools/pcprogress.py passed -fpermissive, which downgrades real errors to
warnings, so it counted 34 units no build would accept: 97.5% reported against
80.3% true. CMakeLists.txt never passed that flag. The flag is gone and the
figures in PCPORT.md and the pc README are corrected.

The biggest remaining class is casting a pointer to U32, which is one decision
nobody has made rather than 23 defects. So the tool reports three lines now:

    compiles            163 / 198 units   82.32%
    pointer width only   23 / 198 units   11.62%
    needs other work     12 / 198 units

-m32 would settle it and cannot be tested here; this container has no multilib.

Fixed this round, all verified codegen-neutral against objects built from clean
source: explicit specializations declared after they were instantiated
(range_limit, auto_tweak::load_param, xListItem<NPCConfig>); 34 redundant
qualifications; u32 vs size_t, one type on the GameCube ABI and two on LP64;
placement operator new, which MSL's <new> does not declare; xVec3Dot and
xVec3Copy, the last PPC assembly in src/SB, keeping ps_sum0's addition order
since float addition is not associative; and __fabs/__fabsf, which glibc
declares and does not define.

Tried and reverted: declaring xFX.cpp's tier_queue member specializations in
xFX.h. That header reaches 71 other units and the declaration lands after
instantiation in all of them -- 191 units became 120.

DOL still 306526d90b48e99894c3138f5fc8f2716d9fecf6; 84 selftest checks pass.

+++ src/SB/Game/zTalkBox.cpp
@@ -1633,7 +1633,10 @@ namespace
 
 } // namespace
 
-void ztalkbox::load(xBase& data, xDynAsset& asset, u32)
+// zTalkBox.h declares this taking size_t. On the GameCube ABI u32 and size_t
+// are both `unsigned long` and the two spellings name one type; on LP64 they
+// do not, and this stops matching its own declaration.
+void ztalkbox::load(xBase& data, xDynAsset& asset, size_t)
 {
     ((ztalkbox&)data).load((const ztalkbox::asset_type&)asset);
 }
